examples/home-energy-advisor/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "home-energy-advisor"
version = "0.1.0"
description = "LangGraph-based energy advisor agent demonstrating ContextForge trajectory evaluation"
requires-python = ">=3.10"
dependencies = [
    # LangGraph and memory support
    "langgraph>=1.0.0",
    "langgraph-checkpoint>=2.0.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    # LangChain
    "langchain>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-core>=1.0.0",
    # Vector store
    "pymilvus>=2.4.0",
    # Utilities
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "contextforge-eval[langgraph]",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "deepeval>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: requires running Ollama instance (deselect with '-m not integration')",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
